# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EAMS API",
    "description": "Events Attendance Monitoring System — attendance capture and administration.\n\n**This document supersedes `docs/api/attendance-contract-handoff.md` as the contract.**\nIt is generated from the source, so the outcome tokens, status codes and payload shapes\nbelow are what the running build actually does.\n\n- **Base path** `/api/v1`. JSON in and out, camelCase field names.\n- **All timestamps are UTC**, ISO 8601, with an explicit `Z`. A bare or offset-bearing\n  local time misjudges the Present/Late boundary by the server's offset.\n- **Errors are RFC 7807** `application/problem+json`. Every body carries a `traceId` to\n  quote back and, on the attendance surface, a stable `code` to branch on — never the\n  human-readable `title` or `detail`, which are reworded freely.\n- **Capture endpoints require a device key** (`DeviceKey` below). Everything else is open\n  for now: authentication for human users is Phase 6, and until it lands this API must\n  stay on a local or trusted network.\n\nSee the `TapOutcomeCode`, `ManualOutcomeCode` and `LiveOutcomeCode` schemas for the frozen\ntoken tables.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/sis/import/upload": {
      "post": {
        "tags": [
          "SisImport"
        ],
        "summary": "Stages a workbook and returns what is in it. Writes nothing to the academic tables.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "description": "The `.xlsx`, as a multipart form file.",
                    "format": "binary"
                  },
                  "termId": {
                    "type": "string",
                    "description": "The term this roster belongs to. <b>Required, and never inferred</b> — see ADR-001 D-5 and\r\n`SisImportBatch.TermId`. Guessing it from the filename or the upload date misfiles an entire\r\nbatch in a way nothing downstream can detect.",
                    "format": "uuid"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                },
                "termId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SisImportPreviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SisImportPreviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SisImportPreviewDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sis/import/{batchId}/run": {
      "post": {
        "tags": [
          "SisImport"
        ],
        "summary": "Runs a staged batch. Idempotent with respect to the database: running the same roster twice\r\nleaves it identical and reports every row `Skipped`.",
        "parameters": [
          {
            "name": "batchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SisImportRunRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SisImportRunRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SisImportRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SisImportBatchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SisImportBatchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SisImportBatchDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sis/import/{batchId}": {
      "get": {
        "tags": [
          "SisImport"
        ],
        "summary": "`GET /sis/import/{batchId}` — one import batch and its counts.",
        "description": "The batch header, not its rows: totals and status, with the staged rows behind\r\n`GET /sis/import/{batchId}/rows`. This is the poll target between uploading a batch and\r\nrunning it, and it is what makes a `404` from `POST .../run` unambiguous — both verbs\r\nagree about which batches exist.",
        "parameters": [
          {
            "name": "batchId",
            "in": "path",
            "description": "The batch.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The batch.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SisImportBatchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SisImportBatchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SisImportBatchDto"
                }
              }
            }
          },
          "404": {
            "description": "No such batch.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sis/import/{batchId}/rows": {
      "get": {
        "tags": [
          "SisImport"
        ],
        "summary": "A batch's staged rows, optionally narrowed to one outcome —\r\n`?result=Failed` is the query an operator runs after every import.",
        "parameters": [
          {
            "name": "batchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "result",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SisImportRowDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SisImportRowDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SisImportRowDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "traceId": {
            "type": "string",
            "description": "**On every error body this API produces.** The correlation handle: the distributed\ntrace id when tracing is on, the per-connection request id otherwise.\n\nQuote it back when reporting a problem — it is the only value that finds the one log\nline explaining an otherwise opaque failure.",
            "nullable": true
          },
          "code": {
            "type": "string",
            "description": "The stable, machine-readable outcome token. **Branch on this, never on `title` or\n`detail`** — those are prose written for a person and are reworded freely.\n\nPresent on the attendance, students and devices surfaces, and on authentication and\nrate-limit refusals. Absent on the events write surface. See the `TapOutcomeCode`,\n`ManualOutcomeCode` and `LiveOutcomeCode` schemas for the frozen token tables.\n\nOn the attendance surface it is the **same field, with the same values**, that a\nsuccess body carries — one accessor works across success and failure alike.",
            "nullable": true
          },
          "serverTime": {
            "type": "string",
            "description": "The server's UTC clock when this response was produced. Present on tap and\nmanual-override failures.\n\nIt is on rejections deliberately, and it matters most on `TappedAtOutOfRange`: that\nresponse tells a device its clock is wrong, so it has to say what the right one is in\nthe same body. Compute an offset from it and apply that before enqueueing.",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "SisImportBatchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "termCode": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "sourceSheetName": {
            "type": "string",
            "nullable": true
          },
          "fileHash": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "totalRows": {
            "type": "integer",
            "format": "int32"
          },
          "insertedRows": {
            "type": "integer",
            "format": "int32"
          },
          "updatedRows": {
            "type": "integer",
            "format": "int32"
          },
          "failedRows": {
            "type": "integer",
            "format": "int32"
          },
          "skippedRows": {
            "type": "integer",
            "format": "int32"
          },
          "warningRows": {
            "type": "integer",
            "format": "int32"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "countersReconcile": {
            "type": "boolean",
            "description": "Whether the counters add up. Exposed rather than left for the caller to compute because it is\r\nthe one check an operator runs, and a UI that has to derive it will eventually derive it wrong.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "A batch's summary. The five outcome counters are the reconciliation ADR-001 D-5 exists to make\r\npossible: `InsertedRows + UpdatedRows + FailedRows + SkippedRows == TotalRows`, always.\r\nEAMS.Application.Dtos.SisImportBatchDto.WarningRows is orthogonal — a warned row is already counted in one of the four."
      },
      "SisImportPreviewDto": {
        "type": "object",
        "properties": {
          "batch": {
            "$ref": "#/components/schemas/SisImportBatchDto"
          },
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "distinctStudents": {
            "type": "integer",
            "format": "int32"
          },
          "distinctColleges": {
            "type": "integer",
            "format": "int32"
          },
          "distinctPrograms": {
            "type": "integer",
            "format": "int32"
          },
          "distinctCourses": {
            "type": "integer",
            "format": "int32"
          },
          "distinctSections": {
            "type": "integer",
            "format": "int32"
          },
          "distinctInstructors": {
            "type": "integer",
            "format": "int32"
          },
          "blankSectionRows": {
            "type": "integer",
            "format": "int32"
          },
          "placeholderInstructorRows": {
            "type": "integer",
            "format": "int32"
          },
          "sampleRows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SisImportRowDto"
            }
          }
        },
        "additionalProperties": false,
        "description": "What an upload found, before anything is written to the academic tables.\r\n            \r\n\r\n\r\nThe distinct counts are the preview's whole purpose: they are what an operator compares against what\r\nthey expect the file to contain, and they are cheap because they come from the staged rows rather\r\nthan from a trial run. A file with one course in it is a wrong file, and this is where that is\r\nvisible — before the run, not after.\r\n"
      },
      "SisImportRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "rowNumber": {
            "type": "integer",
            "format": "int32"
          },
          "result": {
            "type": "string"
          },
          "skipReason": {
            "type": "string",
            "nullable": true
          },
          "warningCode": {
            "type": "string",
            "nullable": true
          },
          "warningMessage": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "studentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "rawData": {
            "type": "string",
            "nullable": true
          },
          "entities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SisImportRowEntityDto"
            }
          }
        },
        "additionalProperties": false,
        "description": "One staged source row and what became of it."
      },
      "SisImportRowEntityDto": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "One entity a row touched. The answer to \"row 214 says Skipped — against what?\", which §4.12's single\r\nnullable `StudentId` could not give for a source whose grain spans six entities."
      },
      "SisImportRunRequest": {
        "type": "object",
        "properties": {
          "termId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "The body of `POST /sis/import/{batchId}/run`.\r\n            \r\n\r\n<b>The term is required here even though the batch already carries one.</b> It is a confirmation,\r\nand it must match or the run is refused. ADR-001 D-5 names the failure this guards: a wrong term\r\nselection misfiles an entire batch, is invisible afterwards (every downstream query is term-scoped,\r\nso the data looks fine — it is simply in the wrong year), and is only recoverable by hand. Asking\r\ntwice, at upload and at the moment of writing, costs one field and turns the mistake into a 409.\r\n"
      },
...
    },
    "securitySchemes": {
      "DeviceKey": {
        "type": "apiKey",
        "description": "The kiosk / capture credential, scoped to `attendance.capture` and nothing else.\n\nSend the **entire header value including the scheme word**:\n\n```\nAuthorization: DeviceKey eams_dk_<keyId>_<secret>\n```\n\nIssued once by `POST /api/v1/devices` and never retrievable afterwards — the server stores\nonly a hash. Store it in the platform keystore (`expo-secure-store` / Android Keystore),\nnever in `AsyncStorage`. Lose it and rotate with `POST /api/v1/devices/{id}/regenerate-key`.\n\nCase is significant and lower-case is the only accepted form for the token itself.\n\nFailures: `401 DeviceKeyMissing` / `DeviceKeyMalformed` / `DeviceKeyInvalid`,\n`403 DeviceKeyRevoked` / `DeviceInactive`, `429 RateLimited` (honour `Retry-After`).\n`DeviceKeyInvalid` deliberately does not distinguish an unknown key id from a wrong secret.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}